moved InputParameters struct to Parameter.h
